requirements.txt
python-terraform==0.10.1
pyyaml==6.0.1
click==8.1.7
python-gitlab==3.15.0
checkov==3.1.9
tqdm==4.66.1
colorama==0.4.6
pytest==7.4.3
pylint==3.0.2